Partially revert a24d5f8a901029be56a78214ce84845d90cdcb5d for OxygenOS
This causes a regression with audio recording in the camera app because
it cannot find the right topology (as verified with logging). This does
not appear to be a problem for custom ROMs at this time; if it turns out
to be later, all changes to this file by a24d5f8a901 should be reverted.

Ultimately, I believe that this should be fixed with an updated audio
mixer file but that is on OnePlus to fix since it is in userspace.

If you are looking to pick this and not using my unified custom
ROM/OxygenOS branch strategy, feel free to use the patch files below
depending on if you have merged LA.UM.6.4.r1-06900-8x98.0 or not.

Pre-LA.UM.6.4.r1-06900-8x98.0 (current OxygenOS Open Beta 6/4 source):

curl https://raw.githubusercontent.com/nathanchance/patches/master/op5/pre-LA.UM.6.4.r1-06900-8x98.0/0001-Revert-ASoC-msm-qdsp6v2-Add-support-for-new-ADM-LSM-.patch | git am -3

Post-LA.UM.6.4.r1-06900-8x98.0:

curl https://raw.githubusercontent.com/nathanchance/patches/master/op5/post-LA.UM.6.4.r1-06900-8x98.0/0001-Revert-ASoC-msm-qdsp6v2-Add-support-for-new-ADM-LSM-.patch | git am -3

Fixes: a24d5f8a9010 ("ASoC: msm: qdsp6v2: Add support for new ADM LSM cal types")
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>